WIP(*): updateing nightwatch to work with local tests
